+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 #### Master
 
+   * New: DelegatingIterable now includes a real implementation of
+     `followedBy`, and accepts the `orElse` parameter on `singleWhere`.
+   * New: DelegatingList now includes real implementations of `operator +`,
+     `indexWhere`, and `lastIndexWhere`.
    * New: The map returned by `Multimap.asMap()` now includes real
      implementations of `get entries` and `removeWhere`. This class also has
      "real" implementations of `addEntries`, `map`, `update`, and `updateAll`,
@@ -53,6 +57,9 @@
      `extent`, `forEachAsync`, `max`, `merge`, `min`, `reduceAsync`, and `zip`
      are now type parameterized. Depending on the inferred value of each type
      parameter, the return type of each function may change in existing code.
+   * BREAKING CHANGE: `Optional`'s `==` operator now takes into account `T`,
+     the type of the value. This changes, e.g. `Optional<int>.absent()` to no
+     longer be equal to `Optional<String>.absent()`.
    * Deprecated: `reverse` in the `strings` library. No replacement is
      provided.
    * Deprecated: `createTimer`, `createTimerPeriodic` in the `async` library.