+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+#### 0.28.0+1 - 2018-03-22
+
+   * Remove use of `Maps.mapToString` in `LruMap`.
+   * Add @visibleForTesting annotation in `AvlTreeSet`.
+
 #### 0.28.0 - 2018-01-19
 
    * BREAKING CHANGE: The signature of `MultiMap`'s `update` stub has changed